Carousel separated from Landing screen and added to new Component
